fix autotools v2 test
